+++ node/src/components/rpc_server/rpcs.rs
@@ -4,6 +4,7 @@
 
 pub mod account;
 pub mod chain;
+mod common;
 pub mod docs;
 pub mod info;
 pub mod state;
@@ -27,6 +28,7 @@ use casper_types::ProtocolVersion;
 
 use super::{ReactorEventT, RpcRequest};
 use crate::effect::EffectBuilder;
+pub use common::ErrorData;
 use docs::DocExample;
 
 /// The URL path.
@@ -49,6 +51,7 @@ enum ErrorCode {
     NoSuchAccount = -32009,
     FailedToGetDictionaryURef = -32010,
     FailedToGetTrie = -32011,
+    NoSuchStateRoot = -32012,
     // Same error code as warp_json INTERNAL_ERROR.
     InternalError = -32063,
 }
@@ -267,66 +270,4 @@ pub(super) trait RpcWithOptionalParamsExt: RpcWithOptionalParams {
         maybe_params: Option<Self::OptionalRequestParams>,
         api_version: ProtocolVersion,
     ) -> BoxFuture<'static, Result<Response<Body>, Error>>;
-}
-
-mod common {
-    use std::convert::TryFrom;
-
-    use once_cell::sync::Lazy;
-
-    use casper_execution_engine::core::engine_state::{self, QueryResult};
-    use casper_types::bytesrepr::ToBytes;
-
-    use super::ErrorCode;
-    use crate::types::json_compatibility::StoredValue;
-
-    pub(super) static MERKLE_PROOF: Lazy<String> = Lazy::new(|| {
-        String::from(
-            "01000000006ef2e0949ac76e55812421f755abe129b6244fe7168b77f47a72536147614625016ef2e0949ac76e\
-        55812421f755abe129b6244fe7168b77f47a72536147614625000000003529cde5c621f857f75f3810611eb4af3\
-        f998caaa9d4a3413cf799f99c67db0307010000006ef2e0949ac76e55812421f755abe129b6244fe7168b77f47a\
-        7253614761462501010102000000006e06000000000074769d28aac597a36a03a932d4b43e4f10bf0403ee5c41d\
-        d035102553f5773631200b9e173e8f05361b681513c14e25e3138639eb03232581db7557c9e8dbbc83ce9450022\
-        6a9a7fe4f2b7b88d5103a4fc7400f02bf89c860c9ccdd56951a2afe9be0e0267006d820fb5676eb2960e15722f7\
-        725f3f8f41030078f8b2e44bf0dc03f71b176d6e800dc5ae9805068c5be6da1a90b2528ee85db0609cc0fb4bd60\
-        bbd559f497a98b67f500e1e3e846592f4918234647fca39830b7e1e6ad6f5b7a99b39af823d82ba1873d0000030\
-        00000010186ff500f287e9b53f823ae1582b1fa429dfede28015125fd233a31ca04d5012002015cc42669a55467\
-        a1fdf49750772bfc1aed59b9b085558eb81510e9b015a7c83b0301e3cf4a34b1db6bfa58808b686cb8fe21ebe0c\
-        1bcbcee522649d2b135fe510fe3")
-    });
-
-    // Extract the EE `(StoredValue, Vec<TrieMerkleProof<Key, StoredValue>>)` from the result.
-    pub(super) fn extract_query_result(
-        query_result: Result<QueryResult, engine_state::Error>,
-    ) -> Result<(StoredValue, Vec<u8>), (ErrorCode, String)> {
-        let (value, proof) = match query_result {
-            Ok(QueryResult::Success { value, proofs }) => (value, proofs),
-            Ok(query_result) => {
-                let error_msg = format!("state query failed: {:?}", query_result);
-                return Err((ErrorCode::QueryFailed, error_msg));
-            }
-            Err(error) => {
-                let error_msg = format!("state query failed to execute: {:?}", error);
-                return Err((ErrorCode::QueryFailedToExecute, error_msg));
-            }
-        };
-
-        let value_compat = match StoredValue::try_from(*value) {
-            Ok(value_compat) => value_compat,
-            Err(error) => {
-                let error_msg = format!("failed to encode stored value: {:?}", error);
-                return Err((ErrorCode::QueryFailed, error_msg));
-            }
-        };
-
-        let proof_bytes = match proof.to_bytes() {
-            Ok(proof_bytes) => proof_bytes,
-            Err(error) => {
-                let error_msg = format!("failed to encode stored value: {:?}", error);
-                return Err((ErrorCode::QueryFailed, error_msg));
-            }
-        };
-
-        Ok((value_compat, proof_bytes))
-    }
 }