+++ node/src/effect.rs
@@ -1019,6 +1019,19 @@ impl<REv> EffectBuilder<REv> {
             .await
     }
 
+    /// Requests the lowest height from which we have an unbroken chain of stored blocks (not just
+    /// headers) ending in the highest stored block.
+    pub(crate) async fn get_lowest_contiguous_block_height_from_storage(self) -> u64
+    where
+        REv: From<StorageRequest>,
+    {
+        self.make_request(
+            |responder| StorageRequest::GetLowestContiguousBlockHeight { responder },
+            QueueKind::Regular,
+        )
+        .await
+    }
+
     /// Get a trie or chunk by its ID.
     pub(crate) async fn get_trie(
         self,